implement movie detail UI part 1

+++ app/src/main/java/com/snad/spotlight/persistence/LibraryMovieDao.kt
@@ -19,5 +19,5 @@ interface LibraryMovieDao {
     fun getAllMovies(): Flow<List<LibraryMovie>>
 
     @Query("SELECT * FROM movies WHERE id LIKE :id")
-    suspend fun getMovieById(id: Long): LibraryMovie?
+    suspend fun getMovieById(id: Int): LibraryMovie?
 }
